PRODUCT: Update roadmap
I've updated the roadmap to reflect what we have planned for this year(they are in the order I think we will be doing them!).

Removed items does not mean they will not be done but rather that they are not a priority at the moment.

Reduced the scope of some items to make them more achievable
and so that we can check more often our roadmap to see if we are going in a good direction.